project_2/pyproject.toml
[project]
name = "project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
